--- src/terra-dev-site/component/Themed.clinical-lowlight-theme.module.scss
@@ -1,5 +1,0 @@
-:local {
-  .clinical-lowlight-theme {
-    --terra-dev-site-themed-block-background-color: #FFA500;
-  }
-}
