--- src/terra-dev-site/component/Themed.module.scss
@@ -1,18 +1,0 @@
-//Themes
-@import './Themed.clinical-lowlight-theme.module';
-@import './Themed.orion-fusion-theme.module';
-@import './Themed.terra-dev-site-test-theme.module.scss';
-
-:local {
-  .themed {
-    height: 100%;
-    overflow: auto;
-    padding: 24px;
-    position: relative;
-  }
-
-  .themed-block {
-    background-color: var(--terra-dev-site-themed-block-background-color, #0079be);
-    height: 24px
-  }
-}
