--- src/terra-dev-site/component/Themed.orion-fusion-theme.module.scss
@@ -1,5 +1,0 @@
-:local {
-  .orion-fusion-theme {
-    --terra-dev-site-themed-block-background-color: #9400D3;
-  }
-}
