--- src/terra-dev-site/component/Themed.terra-dev-site-test-theme.module.scss
@@ -1,5 +1,0 @@
-:local {
-  .terra-dev-site-test-theme {
-    --terra-dev-site-themed-block-background-color: #228B22;
-  }
-}
