+++ src/terra-dev-site/component/ThemedComponent.clinical-lowlight-theme.module.scss
@@ -1,0 +1,5 @@
+:local {
+  .clinical-lowlight-theme {
+    --terra-dev-site-themed-component-block-background-color: #FFA500;
+  }
+}
