+++ src/terra-dev-site/component/ThemedComponent.orion-fusion-theme.module.scss
@@ -1,0 +1,5 @@
+:local {
+  .orion-fusion-theme {
+    --terra-dev-site-themed-component-block-background-color: #9400D3;
+  }
+}
