+++ src/terra-dev-site/component/ThemedComponent.terra-dev-site-test-theme.module.scss
@@ -1,0 +1,5 @@
+:local {
+  .terra-dev-site-test-theme {
+    --terra-dev-site-themed-component-block-background-color: #228B22;
+  }
+}
