+++ src/terra-dev-site/ff/ima folder/ima nother folder/themed.ff.jsx
@@ -1,1 +1,1 @@
-export { default } from '../../../component/Themed';
+export { default } from '../../../component/ThemedComponent';
