+++ src/terra-dev-site/snt/secondary-scroll-test/multiple-files/example2.snt.jsx
@@ -1,1 +1,1 @@
-export { default } from '../../../component/Themed';
+export { default } from '../../../component/ThemedComponent';
