+++ src/terra-dev-site/snt/secondary-scroll-test/parent-folder/child-folder/example.snt.jsx
@@ -1,1 +1,1 @@
-export { default } from '../../../../component/Themed';
+export { default } from '../../../../component/ThemedComponent';
